requirements.txt
torch
torchvision
torchaudio
pyyaml
tqdm
soundfile